% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.tax.DA.R
\name{add.tax.DA}
\alias{add.tax.DA}
\title{Add tax_table from phyloseq object}
\usage{
add.tax.DA(data, res)
}
\arguments{
\item{data}{phyloseq object}

\item{res}{data.frame with results}
}
\description{
Internal function
}
