% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.adx.R
\name{DA.adx}
\alias{DA.adx}
\title{Aldex t.test and wilcox}
\usage{
DA.adx(data, predictor, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{...}{Additional arguments for the \code{aldex} function}
}
\value{
A data.frame with with results.
}
\description{
Implementation of \code{aldex} for \code{DAtest}
}
\examples{
# Creating random count_table and predictor
set.seed(4)
mat <- matrix(rnbinom(1000, size = 0.1, mu = 500), nrow = 100, ncol = 10)
rownames(mat) <- 1:100
pred <- c(rep("Control", 5), rep("Treatment", 5))

# Running ALDEx2
res <- DA.adx(data = mat, predictor = pred)
}
