% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.aoc.R
\name{DA.aoc}
\alias{DA.aoc}
\title{ANOVA - Multiplicative zero-correction and additive log-ratio normalization.}
\usage{
DA.aoc(
  data,
  predictor,
  covars = NULL,
  p.adj = "fdr",
  delta = 1,
  allResults = FALSE,
  ...
)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{covars}{Either a named list with covariables, OR if \code{data} is a \code{phyloseq} object a character vector with names of the variables in \code{sample_data(data)}}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{delta}{Numeric. Pseudocount for zero-correction. Default 1}

\item{allResults}{If TRUE will return raw results from the \code{aov} function}

\item{...}{Additional arguments for the \code{aov} functions}
}
\value{
A data.frame with with results.
}
\description{
Apply ANOVA on multiple features with one \code{predictor}.
}
\examples{
# Creating random count_table and predictor
set.seed(4)
mat <- matrix(rnbinom(1500, size = 0.1, mu = 500), nrow = 100, ncol = 15)
rownames(mat) <- 1:100
pred <- c(rep("A", 5), rep("B", 5), rep("C", 5))

# Running ANOVA on each feature
res <- DA.aoc(data = mat, predictor = pred)
}
