% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocs.R
\name{DA.drop1}
\alias{DA.drop1}
\title{Run \code{drop1} on all features from \code{DAtest} results with \code{allResults = TRUE}}
\usage{
DA.drop1(results, test = "Chisq", p.adj = "fdr", ...)
}
\arguments{
\item{results}{Output from a \code{DA."test"} function with \code{allResults = TRUE}}

\item{test}{Which test to use to calculate p-values. See \code{drop1} for details. Default "Chisq"}

\item{p.adj}{P-value adjustment method. See \code{p.adjust} for details. Default "fdr"}

\item{...}{Additional arguments for \code{drop1} function}
}
\value{
A data.frame with output from drop1 and adjusted p.values for each predictor and feature
}
\description{
Works on "zpo", "znb", "qpo", "neb", "poi". Non-paired "lrm", "llm", "llm2", "lma", "lmc"
}
\examples{
# Creating random count_table, predictor, and covariate
set.seed(5)
mat <- matrix(rnbinom(1500, size = 0.5, mu = 500), nrow = 100, ncol = 15)
rownames(mat) <- 1:100
pred <- c(rep("A", 5), rep("B", 5), rep("C", 5))
covar <- rnorm(15)

# Running linear model and then drop1 on each feature
res <- DA.lmc(mat, pred, covars = list(Something = covar), allResults = TRUE)
res.drop <- DA.drop1(res) 
}
