% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ere2.R
\name{DA.ere2}
\alias{DA.ere2}
\title{EdgeR exact test - RLE normalization}
\usage{
DA.ere2(data, predictor, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{...}{Additional arguments for the calcNormFactors, estimateCommonDisp, estimateTagwiseDisp and exactTest functions}
}
\description{
Implemented as in:
https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-016-0208-8
}
