% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.kru.R
\name{DA.kru}
\alias{DA.kru}
\title{Kruskal-Wallis test}
\usage{
DA.kru(data, predictor, relative = TRUE, p.adj = "fdr",
  allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a phyloseq object the name of the variable in \code{sample_data(data)} in quotation}

\item{relative}{Logical. Should \code{data} be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{allResults}{If TRUE will return raw results from the \code{kruskal.test} function}

\item{...}{Additional arguments for the \code{kruskal.test} function}
}
\value{
A data.frame with with results.
}
\description{
Apply kruskal-wallis test on multiple features with one \code{predictor}
}
\examples{
# Creating random count_table and predictor
set.seed(4)
mat <- matrix(rnbinom(1500, size = 0.1, mu = 500), nrow = 100, ncol = 15)
rownames(mat) <- 1:100
pred <- c(rep("A", 5), rep("B", 5), rep("C", 5))

# Running Kruskal-Wallis on each feature
res <- DA.kru(data = mat, predictor = pred)
}
