% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.lli.R
\name{DA.lli}
\alias{DA.lli}
\title{LIMMA}
\usage{
DA.lli(data, outcome, paired = NULL, relative = TRUE, p.adj = "fdr",
  delta = 1, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{outcome}{The outcome of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{delta}{Numeric. Pseudocount for log transformation. Default 1}

\item{...}{Additional arguments for the eBayes function}
}
\description{
With log transformation for counts before normalization.
Some implementation is borrowed from:
http://www.biostat.jhsph.edu/~kkammers/software/eupa/R_guide.html
}
