% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.lli2.R
\name{DA.lli2}
\alias{DA.lli2}
\title{LIMMA}
\usage{
DA.lli2(count_table, outcome, paired = NULL, p.adj = "fdr", delta = 0.001,
  ...)
}
\arguments{
\item{count_table}{Matrix or data.frame. Table with taxa/genes/proteins as rows and samples as columns}

\item{outcome}{Factor. The outcome of interest. E.g. case and control}

\item{paired}{Factor. Subject ID for running paired analysis}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{delta}{Numeric. Pseudocount for log transformation. Default 0.001}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}
}
\description{
With log transformation of relative abundances.
Some implementation is borrowed from:
http://www.biostat.jhsph.edu/~kkammers/software/eupa/R_guide.html
}
