% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.lma.R
\name{DA.lma}
\alias{DA.lma}
\title{Linear regression - Multiplicative zero-correction and additive log-ratio normalization.}
\usage{
DA.lma(data, predictor, paired = NULL, covars = NULL, out.all = NULL,
  p.adj = "fdr", delta = 1, allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{covars}{Either a named list with covariables, OR if \code{data} is a \code{phyloseq} object a character vector with names of the variables in \code{sample_data(data)}}

\item{out.all}{If TRUE will output results and p-values from \code{anova}. If FALSE will output results for 2. level of the \code{predictor}. If NULL (default) set as TRUE for multi-class \code{predictor} and FALSE otherwise}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{delta}{Numeric. Pseudocount for zero-correction. Default 1}

\item{allResults}{If TRUE will return raw results from the \code{lm}/\code{lme} function}

\item{...}{Additional arguments for the \code{lm}/\code{lme} functions}
}
\description{
Apply linear regression for multiple features with one \code{predictor}.
Mixed-effect model is used when a \code{paired} argument is included, with the \code{paired} variable as a random intercept.
Note: Last feature in the data is used as reference for the log-ratio transformation.
}
