% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocs.R
\name{DA.lsmeans}
\alias{DA.lsmeans}
\title{Run \code{lsmeans} on all features from \code{DAtest} results with \code{allResults = TRUE}}
\usage{
DA.lsmeans(results, variable = "predictor", predictor = NULL,
  covars = NULL, p.adj = "fdr", ...)
}
\arguments{
\item{results}{Output from a \code{DA."test"} function with \code{allResults = TRUE}}

\item{variable}{Which variable to test. Default predictor. Alternatively, the name of a covar}

\item{predictor}{If results come from a paired "lrm", "llm", "lma", "lmc" or "llm2" supply the original predictor variable in the form of as a vector}

\item{covars}{If results come from a paired "lrm", "lma", "lmc", "llm" or "llm2" supply the original covars in the form of a named list}

\item{p.adj}{P-value adjustment method. See \code{p.adjust} for details}

\item{...}{Additional arguments for \code{lsmeans} function}
}
\value{
A data.frame with output from lsmeans::pairs and adjusted p.values for each predictor and feature
}
\description{
Pairwise testing on predictor and covars. Works on "poi", "neb", "lrm", "lma", "lmc", "llm", "llm2", "qpo", "znb", "zpo".
}
\details{
Require the \code{lsmeans} package
}
