% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ltt.R
\name{DA.ltt}
\alias{DA.ltt}
\title{Log t-test}
\usage{
DA.ltt(otu_table, outcome, delta = 1, testStat = function(case, control) {  
    log((mean(case) + 1)/(mean(control) + 1)) },
  testStat.pair = function(case, control) {     mean(log((case + 1)/(control +
  1))) }, paired = NULL, p.adj)
}
\description{
Log t-test
}
