% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ltt2.R
\name{DA.ltt2}
\alias{DA.ltt2}
\title{Welch t-test}
\usage{
DA.ltt2(count_table, outcome, paired = NULL, p.adj = "fdr", delta = 0.001,
  testStat = function(case, control) {     log((mean(case) + 1)/(mean(control)
  + 1)) }, testStat.pair = function(case, control) {     mean(log((case +
  1)/(control + 1))) }, ...)
}
\arguments{
\item{count_table}{Matrix or data.frame. Table with taxa/genes/proteins as rows and samples as columns}

\item{outcome}{Factor. The outcome of interest. E.g. case and control}

\item{paired}{Factor. Subject ID for running paired analysis}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{delta}{Numeric. Pseudocount for log transformation. Default 0.001}

\item{testStat}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a log fold change: log((mean(case abundances)+1)/(mean(control abundances)+1))}

\item{testStat.pair}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a log fold change: mean(log((case abundances+1)/(control abundances+1)))}

\item{...}{Additional arguments for the t.test function}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}
}
\description{
With log transformed relative abundances
}
