% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.msf.R
\name{DA.msf}
\alias{DA.msf}
\title{MetagenomeSeq Feature model}
\usage{
DA.msf(data, predictor, p.adj = "fdr", allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{allResults}{If TRUE will return raw results from the \code{fitFeatureModel} function}

\item{...}{Additional arguments for the \code{fitFeatureModel} function}
}
\value{
A data.frame with with results.
}
\description{
Implemented as in:
https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-016-0208-8
}
\examples{
# Creating random count_table and predictor
set.seed(4)
mat <- matrix(rnbinom(1000, size = 0.1, mu = 500), nrow = 100, ncol = 10)
rownames(mat) <- 1:100
pred <- c(rep("Control", 5), rep("Treatment", 5))

# Running MetagenomeSeq feature model
res <- DA.msf(data = mat, predictor = pred)
}
