% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.mva.R
\name{DA.mva}
\alias{DA.mva}
\title{Mvabund}
\usage{
DA.mva(data, predictor, paired = NULL, covars = NULL, relative = TRUE,
  p.adj = "fdr", coeff = 2, coeff.ref = 1, resamp = "montecarlo",
  allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{covars}{Either a named list with covariables, OR if \code{data} is a \code{phyloseq} object a character vector with names of the variables in \code{sample_data(data)}}

\item{relative}{Logical. Whether \code{log(librarySize)} should be used as offset. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details. Alternatively, "mva" uses mvabunds adjusted p-values}

\item{coeff}{Integer. The p-value and log2FoldChange will be associated with this coefficient. Default 2, i.e. the 2. level of the \code{predictor}.}

\item{coeff.ref}{Integer. Reference level of the \code{predictor}. Will only affect the log2FC and ordering columns on the output. Default the intercept, = 1}

\item{resamp}{Resample method for estimating p-values. Passed to \code{summary.manyglm}. Default "montecarlo"}

\item{allResults}{If TRUE will return raw results from the \code{mvabund} function}

\item{...}{Additional arguments for the \code{manyglm} and \code{summary.manyglm} functions}
}
\description{
Implementation of mvabund manyglm for \code{DAtest}. With negative binomial family and an offset of log(LibrarySize) when relative = TRUE
}
