% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.pea.R
\name{DA.pea}
\alias{DA.pea}
\title{Pearson Correlation}
\usage{
DA.pea(data, predictor, relative = TRUE, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{relative}{Logical. Should \code{data} be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{...}{Additional arguments for the \code{cor.test} function}
}
\value{
A data.frame with with results.
}
\description{
Apply pearson correlation between multiple features and one \code{predictor}
}
\examples{
# Creating random count_table and predictor
set.seed(4)
mat <- matrix(rnbinom(1000, size = 0.1, mu = 500), nrow = 50, ncol = 20)
rownames(mat) <- 1:50
pred <- rnorm(20)

# Running Pearson correlation on each feature
res <- DA.pea(data = mat, predictor = pred)
}
