% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.phyloseq.R
\name{DA.phyloseq}
\alias{DA.phyloseq}
\title{Extract data from a \code{phyloseq} object to be used in \code{DAtest}}
\usage{
DA.phyloseq(data, predictor, paired = NULL, covars = NULL)
}
\arguments{
\item{data}{A \code{phyloseq} object}

\item{predictor}{The \code{predictor} of interest}

\item{paired}{Factor for paired/blocked experimental designs.}

\item{covars}{A character vector with names of the variables in \code{sample_data(data)}}
}
\description{
Extract data from a \code{phyloseq} object to be used in \code{DAtest}
}
