% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.poi.R
\name{DA.poi}
\alias{DA.poi}
\title{Poisson glm}
\usage{
DA.poi(
  data,
  predictor,
  paired = NULL,
  covars = NULL,
  relative = TRUE,
  out.all = NULL,
  p.adj = "fdr",
  coeff = 2,
  coeff.ref = 1,
  allResults = FALSE,
  ...
)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{covars}{Either a named list with covariables, OR if \code{data} is a \code{phyloseq} object a character vector with names of the variables in \code{sample_data(data)}}

\item{relative}{Logical. Whether \code{log(librarySize)} should be used as offset. Default TRUE}

\item{out.all}{If TRUE will output results and p-values from \code{anova} (\code{drop1} if \code{paired != NULL}). If false will output results for 2. level of the \code{predictor}. If NULL (default) set as TRUE for multi-class \code{predictor} and FALSE otherwise}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{coeff}{Integer. The p-value and log2FoldChange will be associated with this coefficient. Default 2, i.e. the 2. level of the \code{predictor}.}

\item{coeff.ref}{Integer. Reference level of the \code{predictor}. Will only affect the log2FC and ordering columns on the output. Default the intercept, = 1}

\item{allResults}{If TRUE will return raw results from the \code{glm}/\code{glmer} function}

\item{...}{Additional arguments for the \code{glm}/\code{glmer} functions}
}
\value{
A data.frame with with results.
}
\description{
Apply poisson generalized linear models on multiple features with one \code{predictor}
With \code{log(librarySize)} as offset.
Mixed-effect model is used when a \code{paired} argument is included, with the \code{paired} variable as a random intercept.
}
\examples{
# Creating random count_table and predictor
set.seed(4)
mat <- matrix(rnbinom(1000, size = 0.1, mu = 500), nrow = 100, ncol = 10)
rownames(mat) <- 1:100
pred <- c(rep("Control", 5), rep("Treatment", 5))

# Running Poisson regression on each feature
res <- DA.poi(data = mat, predictor = pred)
}
