% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.qua.R
\name{DA.qua}
\alias{DA.qua}
\title{Quade test}
\usage{
DA.qua(data, predictor, paired = NULL, relative = TRUE, p.adj = "fdr",
  allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{allResults}{If TRUE will return raw results from the quade.test function}

\item{...}{Additional arguments for the quade.test function}
}
\description{
Quade test
}
