% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.rai.R
\name{DA.rai}
\alias{DA.rai}
\title{RAIDA}
\usage{
DA.rai(count_table, outcome, p.adj = "fdr", ...)
}
\arguments{
\item{count_table}{Matrix or data.frame. Table with taxa/genes/proteins as rows and samples as columns}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{...}{Additional arguments for the raida function}

\item{predictor}{Factor. The outcome of interest. E.g. case and control}
}
\description{
RAIDA
}
