% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.tta.R
\name{DA.tta}
\alias{DA.tta}
\title{Welch t-test - Multiplicative zero-correction and additive log-ratio normalization.}
\usage{
DA.tta(data, predictor, paired = NULL, p.adj = "fdr", delta = 1,
  testStat = function(case, control) {     mean(case) - mean(control) },
  testStat.pair = function(case, control) {     mean(case - control) },
  allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{delta}{Numeric. Pseudocount for zero-correction.}

\item{testStat}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a simple difference: \code{mean(case abundances)-mean(control abundances)}}

\item{testStat.pair}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a simple difference: \code{mean(case abundances-control abundances)}}

\item{allResults}{If TRUE will return raw results from the \code{t.test} function}

\item{...}{Additional arguments for the \code{t.test} function}
}
\value{
A data.frame with with results.
}
\description{
Apply welch t-test to multiple features with one \code{predictor}
}
\details{
Note: Last feature in the data is used as reference for the log-ratio transformation.
}
