% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ttt.R
\name{DA.ttt}
\alias{DA.ttt}
\title{Welch t-test}
\usage{
DA.ttt(data, outcome, paired = NULL, relative = TRUE, p.adj = "fdr",
  testStat = function(case, control) {     log((mean(case) + 1)/(mean(control)
  + 1)) }, testStat.pair = function(case, control) {     mean(log((case +
  1)/(control + 1))) }, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{outcome}{The outcome of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{testStat}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a log fold change: log((mean(case abundances)+1)/(mean(control abundances)+1))}

\item{testStat.pair}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a log fold change: mean(log((case abundances+1)/(control abundances+1)))}

\item{...}{Additional arguments for the t.test function}
}
\description{
Welch t-test
}
