% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurePlot.R
\name{featurePlot}
\alias{featurePlot}
\title{Plot association between abundance of a feature and predictor}
\usage{
featurePlot(data, predictor, paired = NULL, covars = NULL,
  feature = NULL, relative = TRUE, logScale = FALSE, delta = 0.001,
  covar.quant = c(0, 1/3, 2/3, 1))
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples, and there should be rownames}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation.}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data} in quotation.}

\item{covars}{Either a named list with covariates, OR if data is a phyloseq object a character vector with names of the variables in \code{sample_data(data)}}

\item{feature}{Name of feature to plot. Should be in rownames of \code{data} (or \code{taxa_names(data)} if \code{data} is a \code{phyloseq} object)}

\item{relative}{Logical. If TRUE (default) abundances are made relative}

\item{logScale}{Logical. Should abundances be log10-scaled? After normalization if relative is TRUE. Default FALSE}

\item{delta}{Pseudocount for log10 normalization}

\item{covar.quant}{Quantiles for cutting quantitative \code{covars}}
}
\value{
A ggplot
}
\description{
Plot association between abundance of a feature and \code{predictor}, modified if \code{paired} and \code{covars} are available
}
\details{
Boxplot for categorical variables, points and smooth line for quantitative variable.
If a \code{paired} variable is supplied, it is always plotted as points with lines grouped by the \code{paired} variable
If \code{covars} are supplied data is split in facets. Quantitative covars are cut in intervals according to the quantiles given in \code{covar.quant}
}
