% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_alr}
\alias{norm_alr}
\title{Additive log-ratio normalization}
\usage{
norm_alr(x, ref = nrow(x))
}
\arguments{
\item{x}{numeric matrix}

\item{ref}{reference feature}
}
\description{
Additive log-ratio normalization
}
