% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_clr}
\alias{norm_clr}
\title{Centered log-ratio normalization}
\usage{
norm_clr(x)
}
\arguments{
\item{x}{numeric matrix. Samples are columns}
}
\value{
A CLR normalized count_table
}
\description{
Centered log-ratio normalization
}
\examples{
norm_clr(matrix(1:100, nrow = 10, ncol = 10))
}
