% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DA.R
\name{plot.DA}
\alias{plot.DA}
\title{Plotting results from testDA}
\usage{
\method{plot}{DA}(DA, sort = "AUC")
}
\arguments{
\item{DA}{The output from the testDA function}

\item{sort}{Sort methods by c("AUC","FPR","Spike.detect.rate")}
}
\description{
Plotting results from testDA
}
