% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DAPower.R
\name{plot.DAPower}
\alias{plot.DAPower}
\title{Plotting results from \code{powerDA}}
\usage{
\method{plot}{DAPower}(x, ...)
}
\arguments{
\item{x}{The output from the \code{powerDA} function}

\item{...}{Additional arguments for \code{ggdraw}}
}
\description{
Plotting results from \code{powerDA}
}
