% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.DA.R
\name{print.DA}
\alias{print.DA}
\title{Print results from \code{testDA}}
\usage{
\method{print}{DA}(x, ...)
}
\arguments{
\item{x}{The output from the \code{testDA} function}

\item{...}{Additional arguments for \code{print}}
}
\description{
Print results from \code{testDA}
}
