% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneTests.R
\name{pruneTests}
\alias{pruneTests}
\title{Prune tests argument for \code{testDA}}
\usage{
pruneTests(tests, predictor, paired, covars, relative, decimal, zeroes)
}
\arguments{
\item{tests}{A character vector with names of tests}

\item{predictor}{Vector with predictor variable}

\item{paired}{Factor with paired variable}

\item{covars}{Named list with covariables}

\item{relative}{Include tests that work with relative abundances (TRUE) or only absolute abundances (FALSE)}

\item{decimal}{Exclude tests that do not work with decimals (TRUE)}

\item{zeroes}{If FALSE will exclude tests that include zero-inflation}
}
\value{
A character with the tests argument pruned according to the parameters
}
\description{
Prune tests argument for \code{testDA}
}
