% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikein.R
\name{spikein}
\alias{spikein}
\title{Spike-in}
\usage{
spikein(count_table, outcome, spikeMethod = "mult", effectSize = 2, k,
  relative = TRUE)
}
\description{
Modified version of the one from:
https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-016-0208-8
}
