% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.DA.R
\name{summary.DA}
\alias{summary.DA}
\title{Summary of results from testDA}
\usage{
\method{summary}{DA}(object, sort = "AUC", ...)
}
\arguments{
\item{object}{The output from the testDA function}

\item{sort}{Sort methods by c("AUC","FPR","Spike.detect.rate")}

\item{...}{Additional printing arguments}
}
\description{
Summary of results from testDA
}
