% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DAPower.R
\name{summary.DAPower}
\alias{summary.DAPower}
\title{Summary of results from \code{powerDA}}
\usage{
\method{summary}{DAPower}(x, ...)
}
\arguments{
\item{x}{The output from the \code{powerDA} function}

\item{...}{Additional arguments for \code{print}}
}
\description{
Summary of results from \code{powerDA}
}
