% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDA.R
\name{testDA}
\alias{testDA}
\title{Comparing differential abundance methods by FPR and AUC}
\usage{
testDA(otu_table, predictor, R = 5, tests = c("wil", "ttt", "ltt", "ltt2",
  "neb", "erq", "ere", "msf", "zig", "ds2", "per", "bay", "adx"),
  spikeMethod = "mult", effectSize = 2, k = 5, cores = (detectCores() -
  1), rng.seed = 123, delta1 = 1, delta2 = 0.001,
  noOfIterations = 10000, margin = 50, testStat = function(case, control)
  {     log((mean(case) + 1)/(mean(control) + 1)) }, mc.samples = 64)
}
\arguments{
\item{otu_table}{Matrix or data.frame. OTU table with taxa as rows}

\item{predictor}{Factor. The outcome of interest. Should have two levels, e.g. case and control}

\item{R}{Integer. Number of times to run the models. Default 5}

\item{tests}{Character. Which tests to include. Default all}

\item{spikeMethod}{Character. Multiplicative ("mult") or additive ("add") spike-in. Default "mult"}

\item{effectSize}{Integer. The effect size for the spike-ins. Default 2}

\item{k}{Integer. Number of OTUs to spike in each tertile. k=5: 15 OTUs in total. Default 5}

\item{cores}{Integer. Number of cores to use for parallel computing. Default one less than available}

\item{rng.seed}{Numeric. Seed for reproducibility. Default 123}

\item{delta1}{Numeric. The pseudocount for the Log t.test method. Default 1}

\item{delta2}{Numeric. The pseudocount for the Log t.test2 method. Default 0.001}

\item{noOfIterations}{Integer. How many iterations should be run for the permutation test. Default 10000}

\item{margin}{Integer. The margin of when to stop iterating for non-significant OTUs for the permutation test. Default 50}

\item{testStat}{Function. The test statistic function for the permutation test (also in output of ttt, ltt, ltt2 and wil). Should take two vectors as arguments. Default is a log fold change: log((mean(case abundances)+1)/(mean(control abundances)+1))}

\item{mc.samples}{Integer. Monte Carlo samples for ALDEx2. Default 64}
}
\value{
A list of results:
\itemize{
 \item summary - A summary of the results
 \item table - FPR, AUC and spike detection rate for each run
 \item results - A complete list of output from all the methods. Example: Get wilcoxon results from 2. run as such: $results[[2]]["wil"]
}
}
\description{
Calculating false positive rates and AUC for various differential abundance methods
}
\details{
Currently implemented methods:
\itemize{
 \item wil - Wilcoxon Rank Sum on relative abundances
 \item ttt - Welch t.test on relative abundances
 \item ltt - Welch t.test, but reads are first transformed with log(abundance + delta1) then turned into relative abundances
 \item ltt2 - Welch t.test, but with relative abundances transformed with log(relative abundance + delta2)
 \item neb - Negative binomial GLM with log of library size as offset
 \item erq - EdgeR - Quasi likelihood
 \item ere - Edge R - Exact test
 \item msf - MetagenomeSeq feature model
 \item zig - MetagenomeSeq zero-inflated gaussian
 \item ds2 - DESeq2
 \item per - Permutation test with user defined test statistic
 \item bay - baySeq
 \item adx - ALDEx t-test and wilcoxon
}
}
