% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDA.R
\name{allDA}
\alias{allDA}
\title{Run many differential abundance/expression methods}
\usage{
allDA(data, predictor, paired = NULL, covars = NULL, tests = c("neb",
  "per", "adx", "sam", "qua", "fri", "znb", "zpo", "vli", "qpo", "poi",
  "pea", "spe", "wil", "ttt", "ltt", "ltt2", "erq", "ere", "erq2", "ere2",
  "msf", "zig", "ds2", "ds2x", "lim", "lli", "lli2", "aov", "lao", "lao2",
  "kru", "lrm", "llm", "llm2", "tta", "ttc", "aoa", "aoc", "lma", "lmc",
  "lia", "lic"), relative = TRUE, cores = (detectCores() - 1),
  rng.seed = 123, p.adj = "fdr", args = list(), out.all = NULL,
  alpha = 0.1, core.check = TRUE)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples, and there should be rownames}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation. If the \code{predictor} is numeric it will be treated as such in the analyses}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation.}

\item{covars}{Either a named list with covariates, OR if \code{data} is a \code{phyloseq} object a character vector with names of the variables in \code{sample_data(data)}}

\item{tests}{Character. Which tests to include. Default all}

\item{relative}{Logical. TRUE (default) for compositional data. FALSE for absoloute abundances or pre-normalized data.}

\item{cores}{Integer. Number of cores to use for parallel computing. Default one less than available}

\item{rng.seed}{Numeric. Seed for reproducibility. Default 123}

\item{p.adj}{Character. Method for p-value adjustment. See \code{p.adjust} for details. Default "fdr"}

\item{args}{List. A list with lists of arguments passed to the different methods. See details for more.}

\item{out.all}{If TRUE models will output results and p-values from \code{anova}/\code{drop1}. If FALSE will output results for 2. level of the \code{predictor}. If NULL (default) set as TRUE for multi-class \code{predictor} and FALSE otherwise}

\item{alpha}{q-value threshold for calling significance. Default 0.1}

\item{core.check}{If TRUE (default) will make an interactive check that the amount of cores specified are desired. Only if \code{cores>20}. This is to ensure that the function doesn't automatically overloads a server with workers.}
}
\value{
A list of results:
\itemize{
 \item raw - A data.frame with raw p-values from all methods
 \item adj - A data.frame with adjusted p-values from all methods (detection/no-detection from sam)
 \item est - A data.frame with estimates/fold.changes from all relevant methods
 \item details - A dataframe with details from the run
 \item results - A complete list of output from all the methods. Example: Get wilcoxon results from 2. run as such: \code{$results[[2]]["wil"]}
}
}
\description{
Run many differential abundance and expression tests at a time, to easily compare their results
}
\details{
mva and bay are excluded by default, as they often are slow.
}
