% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDA.R
\name{allDA}
\alias{allDA}
\title{Run many differential abundance methods}
\usage{
allDA(data, predictor, paired = NULL, tests = c("spe", "per", "bay", "adx",
  "wil", "ttt", "ltt", "ltt2", "neb", "erq", "ere", "msf", "zig", "ds2", "lim",
  "aov", "lao", "lao2", "kru", "lrm", "llm", "llm2", "rai"), relative = TRUE,
  cores = (detectCores() - 1), rng.seed = 123, p.adj = "fdr",
  args = list(), verbose = FALSE)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation. If the predictor is numeric it will be treated as such in the analyses}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation. Only for "per", "ttt", "ltt", "ltt2", "neb", "wil", "erq", "ds2", "lrm", "llm", "llm2", "lim", "lli" and "lli2"}

\item{tests}{Character. Which tests to include. Default all (See below for details)}

\item{relative}{Logical. Should abundances be made relative? Only has effect for "ttt", "ltt", "wil", "per", "aov", "lao", "kru", "lim", "lli", "lrm", "llm" and "spe". Default TRUE}

\item{cores}{Integer. Number of cores to use for parallel computing. Default one less than available}

\item{rng.seed}{Numeric. Seed for reproducibility. Default 123}

\item{p.adj}{Character. Method for pvalue adjustment. Default "fdr"}

\item{args}{List. A list with lists of arguments passed to the different methods. See details for more.}

\item{verbose}{Logical. Print information during run}
}
\value{
A list of results:
\itemize{
 \item table - Summary of results
 \item results - A complete list of output from all the methods. Example: Get wilcoxon results from 2. run as such: $results[[2]]["wil"]
}
}
\description{
Run many differential abundance tests at a time
}
\details{
Currently implemented methods:
\itemize{
 \item per - Permutation test with user defined test statistic
 \item bay - baySeq
 \item adx - ALDEx t-test and wilcoxon
 \item wil - Wilcoxon Rank Sum on relative abundances
 \item ttt - Welch t.test on relative abundances
 \item ltt - Welch t.test, but reads are first transformed with log(abundance + delta1) then turned into relative abundances
 \item ltt2 - Welch t.test, but with relative abundances transformed with log(relative abundance + delta2)
 \item neb - Negative binomial GLM with log of library size as offset
 \item erq - EdgeR - Quasi likelihood
 \item ere - EdgeR - Exact test
 \item msf - MetagenomeSeq feature model
 \item zig - MetagenomeSeq zero-inflated gaussian
 \item ds2 - DESeq2
 \item lim - LIMMA. Moderated linear models based on emperical bayes
 \item kru - Kruskal-Wallis on relative abundances
 \item aov - ANOVA on relative abundances
 \item lao - ANOVA, but reads are first transformed with log(abundance + delta1) then turned into relative abundances
 \item lao2 - ANOVA, but with relative abundances transformed with log(relative abundance + delta2)
 \item lrm - Linear regression on relative abundances
 \item llm - Linear regression, but reads are first transformed with log(abundance + delta1) then turned into relative abundances
 \item llm2 - Linear regression, but with relative abundances transformed with log(relative abundance + delta2)
 \item rai - RAIDA
 \item spe - Spearman correlation
}

Additional arguments can be passed to the internal functions with the "args" argument. 
It should be structured as a list with elements named by the tests: 
E.g. passing to the DA.per function that it should only run 1000 iterations: args = list(per=list(noOfIterations=1000)).
Include that the log t.test should use a pseudocount of 0.1: args = list(per=list(noOfIterations=1000), ltt=list(delta=0.1)). 
Additional arguments are simply seperated by commas.

Below is an overview of which functions get the arguments that are passed to a specific test
\itemize{
 \item per - Passed to DA.per
 \item bay - Passed to getPriors and getLikelihoods
 \item adx - Passed to aldex
 \item wil - Passed to wilcox.test and DA.wil
 \item ttt - Passed to t.test and DA.ttt
 \item ltt - Passed to t.test and DA.ltt
 \item ltt2 - Passed to t.test and DA.ltt2
 \item neb - Passed to glm.nb and glmer.nb
 \item erq - Passed to exactTest
 \item ere - Passed to glmQLFit
 \item msf - Passed to fitFeatureModel
 \item zig - Passed to fitZig
 \item ds2 - Passed to DESeq
 \item lim - Passed to eBayes
 \item lli - Passed to eBayes
 \item lli2 - Passed to eBayes
 \item kru - Passed to kruskal.test
 \item aov - Passed to aov
 \item lao - Passed to aov
 \item lao2 - Passed to aov
 \item lrm - Passed to lm and lme
 \item llm - Passed to lm and lme
 \item llm2 - Passed to lm and lme
 \item rai - Passed to raida
 \item spe - Passed to cor.test
}
}
