% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.adx.R
\name{DA.adx}
\alias{DA.adx}
\title{Aldex t.test and wilcox}
\usage{
DA.adx(data, predictor, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{...}{Additional arguments for the aldex function}
}
\description{
Aldex t.test and wilcox
}
