% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.adx.R
\name{DA.adx}
\alias{DA.adx}
\title{Aldex t.test and wilcox}
\usage{
DA.adx(data, predictor, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{...}{Additional arguments for the \code{aldex} function}
}
\description{
Implementation of \code{aldex} for \code{DAtest}
}
