% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.anc.R
\name{DA.anc}
\alias{DA.anc}
\title{ANCOM}
\usage{
DA.anc(count_table, outcome, ...)
}
\arguments{
\item{count_table}{Matrix or data.frame. Table with taxa/genes/proteins as rows and samples as columns}

\item{outcome}{Factor. The outcome of interest. E.g. case and control}

\item{...}{Additional arguments for the ANCOM function}
}
\description{
ANCOM
}
