% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.anc.R
\name{DA.anc}
\alias{DA.anc}
\title{ANCOM}
\usage{
DA.anc(data, predictor, paired = NULL, allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{allResults}{If TRUE will return raw results from the ANCOM function}

\item{...}{Additional arguments for the ANCOM function}
}
\description{
ANCOM
}
