% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocs.R
\name{DA.anova}
\alias{DA.anova}
\title{Run anova on all features from DAtest results with allResults = TRUE}
\usage{
DA.anova(results, p.adj = "fdr", ...)
}
\arguments{
\item{results}{Output from a DA."test" function with allResults = TRUE}

\item{p.adj}{P-value adjustment method. See ?p.adjust for details}

\item{...}{Additional arguments for anova function}
}
\description{
Works on "lrm", "llm", "llm2". Non-paired "neb"
}
