% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.bay.R
\name{DA.bay}
\alias{DA.bay}
\title{baySeq}
\usage{
DA.bay(data, predictor, allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{allResults}{If TRUE will return raw results from the \code{getLikelihoods} function}

\item{...}{Additional arguments to the \code{getPriors.NB} and \code{getLikelihoods} functions}
}
\value{
A data.frame with with results.
}
\description{
Implementation of baySeq for \code{DAtest}
}
