% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocs.R
\name{DA.drop1}
\alias{DA.drop1}
\title{Run drop1 on all features from DAtest results with allResults = TRUE}
\usage{
DA.drop1(results, test = "Chisq", p.adj = "fdr", ...)
}
\arguments{
\item{results}{Output from a DA."test" function with allResults = TRUE}

\item{test}{Which test to use to calculate p-values. See ?drop1 for details}

\item{p.adj}{P-value adjustment method. See ?p.adjust for details}

\item{...}{Additional arguments for drop1 function}
}
\description{
Works on "zpo", "znb", "qpo", "neb", "poi". Non-paired "lrm", "llm", "llm2"
}
