% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ds2.R
\name{DA.ds2}
\alias{DA.ds2}
\title{DESeq2}
\usage{
DA.ds2(otu_table, outcome, p.adj)
}
\description{
From:
https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-016-0208-8
Manual geometric means calculated to avoid errors, see https://github.com/joey711/phyloseq/issues/387
}
