% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ds2.R
\name{DA.ds2}
\alias{DA.ds2}
\title{DESeq2}
\usage{
DA.ds2(data, outcome, paired = NULL, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{outcome}{The outcome of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{...}{Additional arguments for the DESeq function}
}
\description{
Implemented as in:
https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-016-0208-8.
Manual geometric means calculated to avoid errors, see https://github.com/joey711/phyloseq/issues/387
}
