% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ere2.R
\name{DA.ere2}
\alias{DA.ere2}
\title{EdgeR exact test - RLE normalization}
\usage{
DA.ere2(data, predictor, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{...}{Additional arguments for the \code{calcNormFactors}, \code{estimateCommonDisp}, \code{estimateTagwiseDisp} and \code{exactTest} functions}
}
\description{
Implementation of edgeR exact test for \code{DAtest}
}
