% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ere2.R
\name{DA.ere2}
\alias{DA.ere2}
\title{EdgeR exact test - RLE normalization}
\usage{
DA.ere2(data, predictor, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{...}{Additional arguments for the \code{calcNormFactors}, \code{estimateCommonDisp}, \code{estimateTagwiseDisp} and \code{exactTest} functions}
}
\value{
A data.frame with with results.
}
\description{
Implementation of edgeR exact test for \code{DAtest}
}
\examples{
# Creating random count_table and predictor
set.seed(5)
mat <- matrix(rnbinom(1000, size = 0.1, mu = 500), nrow = 100, ncol = 10)
rownames(mat) <- 1:100
pred <- c(rep("Control", 5), rep("Treatment", 5))

# Running edgeR
res <- DA.ere2(data = mat, predictor = pred)
}
