% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.kru.R
\name{DA.kru}
\alias{DA.kru}
\title{Kruskal-Wallis test}
\usage{
DA.kru(data, predictor, relative = TRUE, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{...}{Additional arguments for the kruskal.test function}
}
\description{
Kruskal-Wallis test
}
