% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.lao2.R
\name{DA.lao2}
\alias{DA.lao2}
\title{ANOVA}
\usage{
DA.lao2(data, predictor, covars = NULL, p.adj = "fdr", delta = 0.001,
  allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{covars}{Either a named list with covariables, OR if data is a phyloseq object a character vector with names of the variables in sample_data(data)}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{delta}{Numeric. Pseudocount for the log transformation. Default 0.001}

\item{allResults}{If TRUE will return raw results from the aov function}

\item{...}{Additional arguments for the aov functions}
}
\description{
With log transformation of relative abundances.
}
