% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.lim.R
\name{DA.lim}
\alias{DA.lim}
\title{LIMMA}
\usage{
DA.lim(count_table, outcome, p.adj, relative, log = FALSE, delta = 1)
}
\description{
Some is borrowed from:
http://www.biostat.jhsph.edu/~kkammers/software/eupa/R_guide.html
}
