% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.lli.R
\name{DA.lli}
\alias{DA.lli}
\title{LIMMA}
\usage{
DA.lli(count_table, outcome, paired = NULL, relative = TRUE,
  p.adj = "fdr", delta = 1, ...)
}
\arguments{
\item{count_table}{Matrix or data.frame. Table with taxa/genes/proteins as rows and samples as columns}

\item{outcome}{Factor. The outcome of interest. E.g. case and control}

\item{paired}{Factor. Subject ID for running paired analysis}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{delta}{Numeric. Pseudocount for log transformation. Default 1}
}
\description{
With log transformation for counts before normalization.
Some implementation is borrowed from:
http://www.biostat.jhsph.edu/~kkammers/software/eupa/R_guide.html
}
