% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.lli2.R
\name{DA.lli2}
\alias{DA.lli2}
\title{LIMMA}
\usage{
DA.lli2(data, predictor, paired = NULL, covars = NULL, out.anova = TRUE,
  p.adj = "fdr", delta = 0.001, allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{covars}{Either a named list with covariables, OR if data is a phyloseq object a character vector with names of the variables in sample_data(data)}

\item{out.anova}{If TRUE will output results from F-tests, if FALSE t-statistic results from 2. level of the predictor.}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{delta}{Numeric. Pseudocount for log transformation. Default 0.001}

\item{allResults}{If TRUE will return raw results from the eBayes function}

\item{...}{Additional arguments for the eBayes and lmFit functions}
}
\description{
With log transformation of relative abundances.
Some implementation is borrowed from:
http://www.biostat.jhsph.edu/~kkammers/software/eupa/R_guide.html
}
