% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.llm.R
\name{DA.llm}
\alias{DA.llm}
\title{Linear regression}
\usage{
DA.llm(data, predictor, paired = NULL, covars = NULL, relative = TRUE,
  out.anova = NULL, p.adj = "fdr", delta = 1, allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{covars}{Either a named list with covariables, OR if data is a phyloseq object a character vector with names of the variables in sample_data(data)}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}

\item{out.anova}{If TRUE will output results and p-values from anova. If false will output results for 2. level of the predictor. If NULL (default) set as TRUE for multi-class predictors and FALSE otherwise}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{delta}{Numeric. Pseudocount for the log transformation. Default 1}

\item{allResults}{If TRUE will return raw results from the lm/lme function}

\item{...}{Additional arguments for the lm/lme functions}
}
\description{
With log transformation of counts before normalization.
Mixed-effect model is used when a paired argument is included, with the paired variable as a random intercept.
}
