% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.llm2.R
\name{DA.llm2}
\alias{DA.llm2}
\title{Linear regression}
\usage{
DA.llm2(data, predictor, paired = NULL, covars = NULL, p.adj = "fdr",
  delta = 0.001, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{covars}{Either a named list with covariables, OR if data is a phyloseq object a character vector with names of the variables in sample_data(data)}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{delta}{Numeric. Pseudocount for the log transformation. Default 0.001}

\item{...}{Additional arguments for the lm/lme functions}
}
\description{
With log transformation of relative abundances.
Mixed-effect model is used when a paired argument is included, with the paired variable as a random intercept.
}
