% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.llm2.R
\name{DA.llm2}
\alias{DA.llm2}
\title{Log linear regression}
\usage{
DA.llm2(data, predictor, paired = NULL, covars = NULL,
  out.all = NULL, p.adj = "fdr", delta = 0.001, allResults = FALSE,
  ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{covars}{Either a named list with covariables, OR if \code{data} is a \code{phyloseq} object a character vector with names of the variables in sample_data(data)}

\item{out.all}{If TRUE will output results and p-values from \code{anova}. If FALSE will output results for 2. level of the \code{predictor}. If NULL (default) set as TRUE for multi-class \code{predictor} and FALSE otherwise}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{delta}{Numeric. Pseudocount for the log transformation. Default 0.001}

\item{allResults}{If TRUE will return raw results from the \code{lm}/\code{lme} function}

\item{...}{Additional arguments for the \code{lm}/\code{lme} functions}
}
\description{
Apply linear regression on multiple features with one \code{predictor}, with log transformation of relative abundances.
Mixed-effect model is used when a \code{paired} argument is included, with the \code{paired} variable as a random intercept.
}
