% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.lrm.R
\name{DA.lrm}
\alias{DA.lrm}
\title{Linear regression}
\usage{
DA.lrm(data, predictor, paired = NULL, covars = NULL, relative = TRUE,
  p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{covars}{Either a named list with covariables, OR if data is a phyloseq object a character vector with names of the variables in sample_data(data)}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{...}{Additional arguments for the lm/lme functions}
}
\description{
Mixed-effect model is a paired argument is present, with the paired variable as random intercept
}
