% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocs.R
\name{DA.lsmeans}
\alias{DA.lsmeans}
\title{Run lsmeans on all features from DAtest results with allResults = TRUE}
\usage{
DA.lsmeans(results, variable = "predictor", predictor = NULL,
  covars = NULL, p.adj = "fdr", ...)
}
\arguments{
\item{results}{Output from a DA."test" function with allResults = TRUE}

\item{variable}{Which variable to test. Default predictor. Alternatively, the name of a covar}

\item{predictor}{If results come from a paired "lrm", "llm" or "llm2" supply the original predictor variable in the form of as a vector}

\item{covars}{If results come from a paired "lrm", "llm" or "llm2" supply the original covars in the form of a named list}

\item{p.adj}{P-value adjustment method. See ?p.adjust for details}

\item{...}{Additional arguments for lsmeans function}
}
\description{
Pairwise testing on predictor and covars. Works on "poi", "neb", "lrm", "llm", "llm2", "qpo", "znb", "zpo".
}
\details{
Require the lsmeans package
}
