% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ltt2.R
\name{DA.ltt2}
\alias{DA.ltt2}
\title{Welch t-test}
\usage{
DA.ltt2(data, predictor, paired = NULL, p.adj = "fdr", delta = 0.001,
  testStat = function(case, control) {     (mean(case)) - (mean(control)) },
  testStat.pair = function(case, control) {     mean(case - control) },
  allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{delta}{Numeric. Pseudocount for log transformation. Default 0.001}

\item{testStat}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a mean(case abundances)-mean(control abundances)}

\item{testStat.pair}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a log fold change: mean(case abundances - control abundances)}

\item{allResults}{If TRUE will return raw results from the t.test function}

\item{...}{Additional arguments for the t.test function}
}
\description{
With log transformed relative abundances
}
