% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ltt2.R
\name{DA.ltt2}
\alias{DA.ltt2}
\title{Log t-test2}
\usage{
DA.ltt2(count_table, outcome, delta = 0.001, testStat = function(case,
  control) {     log((mean(case) + 1)/(mean(control) + 1)) },
  testStat.pair = function(case, control) {     mean(log((case + 1)/(control +
  1))) }, paired = NULL, p.adj)
}
\description{
Log t-test2
}
