% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.msf.R
\name{DA.msf}
\alias{DA.msf}
\title{MetagenomeSeq Feature model}
\usage{
DA.msf(data, predictor, p.adj = "fdr", allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{allResults}{If TRUE will return raw results from the \code{fitFeatureModel} function}

\item{...}{Additional arguments for the \code{fitFeatureModel} function}
}
\value{
A data.frame with with results.
}
\description{
Implemented as in:
https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-016-0208-8
}
