% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.msf.R
\name{DA.msf}
\alias{DA.msf}
\title{MetagenomeSeq Feature model}
\usage{
DA.msf(count_table, outcome, p.adj = "fdr", ...)
}
\arguments{
\item{count_table}{Matrix or data.frame. Table with taxa/genes/proteins as rows and samples as columns}

\item{outcome}{Factor. The outcome of interest. E.g. case and control}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{...}{Additional arguments for the fitFeatureModel function}
}
\description{
Implemented as in:
https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-016-0208-8
}
