% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.neb.R
\name{DA.neb}
\alias{DA.neb}
\title{Negative binomial glm}
\usage{
DA.neb(data, predictor, paired = NULL, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{...}{Additional arguments for the glm.nb/glmer.nb functions}
}
\description{
With log(librarySize) as offset.
Mixed-effect model is used when a paired argument is included, with the paired variable as a random intercept.
}
