% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.per.R
\name{DA.per}
\alias{DA.per}
\title{Permutation test}
\usage{
DA.per(otu_table, outcome, noOfIterations = 10000,
  seed = as.numeric(Sys.time()), margin = 50, testStat = function(case,
  control) {     log((mean(case) + 1)/(mean(control) + 1)) })
}
\description{
Modified version of the one from:
https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-016-0208-8
Pvalues are now two-sided, and test statistic is log fold change
}
