% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.qpo.R
\name{DA.qpo}
\alias{DA.qpo}
\title{Quasi-poisson glm}
\usage{
DA.qpo(data, predictor, covars = NULL, relative = TRUE, out.all = NULL,
  p.adj = "fdr", coeff = 2, coeff.ref = 1, allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{covars}{Either a named list with covariables, OR if \code{data} is a \code{phyloseq} object a character vector with names of the variables in \code{sample_data(data)}}

\item{relative}{Logical. Whether \code{log(librarySize)} should be used as offset. Default TRUE}

\item{out.all}{If TRUE will output results and p-values from \code{anova}. If FALSE will output results for 2. level of the \code{predictor}. If NULL (default) set as TRUE for multi-class \code{predictor} and FALSE otherwise}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{coeff}{Integer. The p-value and log2FoldChange will be associated with this coefficient. Default 2, i.e. the 2. level of the \code{predictor}.}

\item{coeff.ref}{Integer. Reference level of the \code{predictor}. Will only affect the log2FC and ordering columns on the output. Default the intercept, = 1}

\item{allResults}{If TRUE will return raw results from the \code{glm} function}

\item{...}{Additional arguments for the \code{glm} functions}
}
\description{
Apply quasi-poisson generalized linear model for multiple features with one \code{predictor}
With \code{log(librarySize)} as offset if \code{relative=TRUE}.
}
