% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.rai.R
\name{DA.rai}
\alias{DA.rai}
\title{RAIDA}
\usage{
DA.rai(data, predictor, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{...}{Additional arguments for the \code{raida} function}
}
\description{
Implementation of \code{raida} for \code{DAtest}
}
