% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.spe.R
\name{DA.spe}
\alias{DA.spe}
\title{Spearman's Rank Correlation}
\usage{
DA.spe(data, predictor, relative = TRUE, p.adj = "fdr", ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{relative}{Logical. Should \code{data} be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{...}{Additional arguments for the \code{cor.test} function}
}
\description{
Apply Spearman correlation to multiple features with one \code{predictor}
}
