% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.ttt.R
\name{DA.ttt}
\alias{DA.ttt}
\title{t-test}
\usage{
DA.ttt(count_table, outcome, testStat = function(case, control) {    
  log((mean(case) + 1)/(mean(control) + 1)) }, testStat.pair = function(case,
  control) {     mean(log((case + 1)/(control + 1))) }, paired = NULL, p.adj,
  relative = TRUE)
}
\description{
t-test
}
