% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocs.R
\name{DA.TukeyHSD}
\alias{DA.TukeyHSD}
\title{Run \code{TukeyHSD} on all features from \code{DAtest} results with \code{allResults = TRUE}}
\usage{
DA.TukeyHSD(results, variable = "predictor", p.adj = "fdr", ...)
}
\arguments{
\item{results}{Output from a \code{DA."test"} function with \code{allResults = TRUE}}

\item{variable}{Which variable to test. Default predictor. Alternatively, the name of a covar}

\item{p.adj}{P-value adjustment method. See \code{p.adjust} for details}

\item{...}{Additional arguments for \code{TukeyHSD} function}
}
\value{
A data.frame with output from TukeyHSD and adjusted p.values for each predictor and feature
}
\description{
Works on "aov", "lao", "lao2", "aoc", "aoa"
}
\examples{
# Creating random count_table and predictor
set.seed(5)
mat <- matrix(rnbinom(3000, size = 0.1, mu = 500), nrow = 100, ncol = 30)
rownames(mat) <- 1:100
pred <- c(rep("A", 10), rep("B", 10), rep("C", 10))

# Run anova and then TukeyHSD on each predictor
res <- DA.aov(data = mat, predictor = pred, allResults = TRUE)
res.tuk <- DA.TukeyHSD(res)
}
