% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.vli.R
\name{DA.vli}
\alias{DA.vli}
\title{LIMMA with \code{voom}}
\usage{
DA.vli(
  data,
  predictor,
  paired = NULL,
  covars = NULL,
  out.all = NULL,
  p.adj = "fdr",
  coeff = 2,
  allResults = FALSE,
  ...
)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a \code{phyloseq} object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation}

\item{covars}{Either a named list with covariables, OR if \code{data} is a \code{phyloseq} object a character vector with names of the variables in \code{sample_data(data)}}

\item{out.all}{If TRUE will output results from F-tests, if FALSE t-statistic results from 2. level of the \code{predictor}. If NULL (default) set as TRUE for multi-class predictors and FALSE otherwise}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See \code{p.adjust} for details}

\item{coeff}{Integer. The p-value and log2FoldChange will be associated with this coefficient. Default 2, i.e. the 2. level of the \code{predictor}.}

\item{allResults}{If TRUE will return raw results from the \code{eBayes} function}

\item{...}{Additional arguments for the \code{voom}, \code{eBayes} and \code{lmFit} functions}
}
\value{
A data.frame with with results.
}
\description{
Limma is run after variance stabilization with \code{voom} using \code{edgeR} normalized library sizes
}
\examples{
# Creating random count_table and predictor
set.seed(4)
mat <- matrix(rnbinom(1000, size = 0.1, mu = 500), nrow = 100, ncol = 10)
rownames(mat) <- 1:100
pred <- c(rep("Control", 5), rep("Treatment", 5))

# Running limma
res <- DA.vli(data = mat, predictor = pred)
}
