% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.wil.R
\name{DA.wil}
\alias{DA.wil}
\title{Wilcoxon Rank Sum and Signed Rank Test}
\usage{
DA.wil(data, predictor, paired = NULL, relative = TRUE, p.adj = "fdr",
  testStat = function(case, control) {     log((mean(case) + 1)/(mean(control)
  + 1)) }, testStat.pair = function(case, control) {     mean(log((case +
  1)/(control + 1))) }, allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Factor, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{relative}{Logical. Should count_table be normalized to relative abundances. Default TRUE}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{testStat}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a log fold change: log((mean(case abundances)+1)/(mean(control abundances)+1))}

\item{testStat.pair}{Function. Function for calculating fold change. Should take two vectors as arguments. Default is a log fold change: mean(log((case abundances+1)/(control abundances+1)))}

\item{allResults}{If TRUE will return raw results from the wilcox.test function}

\item{...}{Additional arguments for the wilcox.test function}
}
\description{
Wilcoxon Rank Sum and Signed Rank Test
}
