% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.zig.R
\name{DA.zig}
\alias{DA.zig}
\title{MetageonomeSeq ZIG}
\usage{
DA.zig(data, predictor, paired = NULL, covars = NULL, p.adj = "fdr",
  by = 2, allResults = FALSE, ...)
}
\arguments{
\item{data}{Either a matrix with counts/abundances, OR a phyloseq object. If a matrix/data.frame is provided rows should be taxa/genes/proteins and columns samples}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if data is a phyloseq object the name of the variable in sample_data in quotation}

\item{covars}{Either a named list with covariables, OR if data is a phyloseq object a character vector with names of the variables in sample_data(data)}

\item{p.adj}{Character. P-value adjustment. Default "fdr". See p.adjust for details}

\item{by}{Column number or column name specifying which coefficient or contrast of the linear model is of interest (only for categorical predictors). Default 2}

\item{allResults}{If TRUE will return raw results from the fitZig function}

\item{...}{Additional arguments for the fitZig function}
}
\description{
Implemented as in:
https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-016-0208-8
}
