% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DA.R
\name{plot.DA}
\alias{plot.DA}
\title{Plotting results from testDA}
\usage{
\method{plot}{DA}(x, sort = "AUC", p = FALSE, bins = 20, adj = FALSE,
  ...)
}
\arguments{
\item{x}{The output from the testDA function}

\item{sort}{Sort methods by c("AUC","FPR")}

\item{p}{Logical. Should the p-value distribution be plotted (only p-values from non-spiked features)}

\item{bins}{Integer. Number of bins in p-value histograms}

\item{adj}{Logical. Whether the histograms should show adjusted p-values}

\item{...}{Additional plotting arguments}
}
\description{
Plotting results from testDA
}
