% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.DA.R
\name{print.DA}
\alias{print.DA}
\title{Print results from testDA}
\usage{
\method{print}{DA}(x, ...)
}
\arguments{
\item{x}{The output from the testDA function}

\item{...}{Additional printing arguments}
}
\description{
Print results from testDA
}
