% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.DA.R
\name{summary.DA}
\alias{summary.DA}
\title{Summary and print of results from testDA}
\usage{
\method{summary}{DA}(DA)
}
\arguments{
\item{DA}{The output from the testDA function}
}
\description{
Summary and print of results from testDA
}
