% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.DA.R
\name{summary.DA}
\alias{summary.DA}
\title{Summary of results from testDA}
\usage{
\method{summary}{DA}(DA, sort = "AUC")
}
\arguments{
\item{DA}{The output from the testDA function}

\item{sort}{Sort methods by c("AUC","FPR","Spike.detect.rate")}
}
\description{
Summary of results from testDA
}
