% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.DA.R
\name{summary.DA}
\alias{summary.DA}
\title{Summary of results from \code{testDA}}
\usage{
\method{summary}{DA}(object, sort = "Rank", ...)
}
\arguments{
\item{object}{The output from the \code{testDA} function}

\item{sort}{Sort methods by \code{c("AUC","FPR","Spike.detect.rate","Rank")}}

\item{...}{Additional arguments for \code{print}}
}
\description{
Summary of results from \code{testDA}
}
