% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DA.R
\name{summary.DA}
\alias{summary.DA}
\title{Summary of results from \code{testDA}}
\usage{
\method{summary}{DA}(object, sort = "Score", boot = TRUE,
  prob = c(0.05, 0.95), N = 1000, decimals = 2, ...)
}
\arguments{
\item{object}{The output from the \code{testDA} function}

\item{sort}{Sort methods by \code{c("AUC","FPR","Power","Score")}}

\item{boot}{If TRUE will use bootstrap for confidence limits of the Score, else will compute the limits from the original table. Recommended to be TRUE unless \code{R >= 100} in \code{testDA}}

\item{prob}{Confidence limits for Score. Default \code{90\%} = \code{c(0.05,0.095)}}

\item{N}{Number of bootstraps. Default 1000}

\item{decimals}{Precision in output. Default 2}

\item{...}{Additional arguments for \code{print}}
}
\value{
Prints summary from the DAtest function. '*' next to method, means that it's median score is overlapping with the 90% confidence limits of the best method
}
\description{
Summary of results from \code{testDA}
}
