% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DAPower.R
\name{summary.DAPower}
\alias{summary.DAPower}
\title{Summary of results from \code{powerDA}}
\usage{
\method{summary}{DAPower}(object, decimals = 2, ...)
}
\arguments{
\item{object}{The output from the \code{powerDA} function}

\item{decimals}{Precision in output. Default 2}

\item{...}{Additional arguments for \code{print}}
}
\value{
Prints summary results from the powerDA function
}
\description{
Summary of results from \code{powerDA}
}
