% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDA.R
\name{testDA}
\alias{testDA}
\title{Comparing differential abundance/expression methods by Empirical power and False Discovery Rate}
\usage{
testDA(data, predictor, paired = NULL, covars = NULL, R = 20,
  tests = c("bay", "ds2", "ds2x", "per", "adx", "znb", "zpo", "msf",
  "zig", "erq", "erq2", "neb", "qpo", "poi", "sam", "lrm", "llm", "llm2",
  "lma", "lmc", "ere", "ere2", "pea", "spe", "wil", "kru", "qua", "fri",
  "ttt", "ltt", "ltt2", "tta", "ttc", "aov", "lao", "lao2", "aoa", "aoc",
  "vli", "lim", "lli", "lli2", "lia", "lic"), relative = TRUE,
  effectSize = 5, k = NULL, cores = (detectCores() - 1),
  p.adj = "fdr", args = list(), out.all = NULL, alpha = 0.1,
  core.check = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{Either a data.frame with counts/abundances, OR a \code{phyloseq} object. If a data.frame is provided rows should be taxa/genes/proteins and columns samples, and there should be rownames}

\item{predictor}{The predictor of interest. Either a Factor or Numeric, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation. If the \code{predictor} is numeric it will be treated as such in the analyses}

\item{paired}{For paired/blocked experimental designs. Either a Factor with Subject/Block ID for running paired/blocked analysis, OR if \code{data} is a \code{phyloseq} object the name of the variable in \code{sample_data(data)} in quotation.}

\item{covars}{Either a named list with covariates, OR if \code{data} is a \code{phyloseq} object a character vector with names of the variables in \code{sample_data(data)}}

\item{R}{Integer. Number of times to run the tests. Default 20}

\item{tests}{Character. Which tests to include. Default all}

\item{relative}{Logical. TRUE (default) for compositional data. FALSE for absolute abundances or pre-normalized data.}

\item{effectSize}{Numeric. The effect size for the spike-ins. Default 5}

\item{k}{Vector of length 3. Number of Features to spike in each tertile (lower, mid, upper). E.g. \code{k=c(5,10,15)}: 5 features spiked in low abundance tertile, 10 features spiked in mid abundance tertile and 15 features spiked in high abundance tertile. Default NULL, which will spike 2 percent of the total amount of features in each tertile (a total of 6 percent), but minimum c(5,5,5)}

\item{cores}{Integer. Number of cores to use for parallel computing. Default one less than available. Set to 1 for sequential computing.}

\item{p.adj}{Character. Method for p-value adjustment. See \code{p.adjust} for details. Default "fdr"}

\item{args}{List. A list with lists of arguments passed to the different methods. See details for more.}

\item{out.all}{If TRUE linear models will output results and p-values from \code{anova}/\code{drop1}, ds2/ds2x will run LRT and not Wald test, erq and erq2 will produce one p-value for the predictor, and limma will run F-tests. If FALSE will output results for 2. level of the \code{predictor}. If NULL (default) set as TRUE for multi-class predictors and FALSE otherwise}

\item{alpha}{q-value threshold for determining significance for \code{Power}. Default 0.1}

\item{core.check}{If TRUE will make an interactive check that the amount of cores specified are desired. Only if \code{cores>20}. This is to ensure that the function doesn't automatically overloads a server with workers.}

\item{verbose}{If TRUE will print informative messages}
}
\value{
An object of class \code{DA}, which contains a list of results:
\itemize{
 \item table - FPR, AUC and spike detection rate for each run
 \item results - A complete list of output from all the methods. Example: Get wilcoxon results from 2. run as such: \code{$results[[2]]["wil"]}
 \item details - A dataframe with details from the run
 \item run.times - A dataframe with average run times of the different methods
}
}
\description{
Calculating Power, False Discovery Rates, False Positive Rates and AUC (Area Under the Receiver Operating Characteristic (ROC) Curve) for various differential abundance and expression methods
}
\details{
mva is excluded by default, as it is slow.
}
