% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDA.R
\name{vennDA}
\alias{vennDA}
\title{Plot Venn diagram from allDA object}
\usage{
vennDA(x, tests = NULL, alpha = 0.05, ...)
}
\arguments{
\item{x}{Output from the allDA function}

\item{tests}{Character vector with tests to plot (E.g. c("ttt","adx.t","wil"), see names(x$results)). Default none}

\item{alpha}{Numeric. q-value threshold for significant features. Default 0.05}

\item{...}{Additional arguments for plotting}
}
\value{
Nothing
}
\description{
Plot a Venn (Euler) diagram of features found by different methods.
}
\details{
Require the venneuler package.
}
