% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDA.R
\name{vennDA}
\alias{vennDA}
\title{Plot Venn diagram from \code{allDA} object}
\usage{
vennDA(x, tests = NULL, alpha = 0.05, split = FALSE, output = FALSE,
  pkg = "eulerr", ...)
}
\arguments{
\item{x}{(Required) Output from the \code{allDA} function}

\item{tests}{(Required) Character vector with tests to plot (E.g. \code{c("ttt","adx.t","wil")}, see \code{names(x$results)}). Default none}

\item{alpha}{Numeric. q-value threshold for significant features. Default 0.05}

\item{split}{If TRUE will split diagrams in positive and negative estimates if possible}

\item{output}{If TRUE will return a data.frame instead of a plot}

\item{pkg}{Use either "eulerr" package (default) or "venneuler" for drawing diagrams.}

\item{...}{Additional arguments for plotting}
}
\value{
If output TRUE then a data.frame with Features detected by the different methods
}
\description{
Plot a Venn (Euler) diagram of features found by different methods.
}
\details{
Require the eulerr package unless output is TRUE.
}
